gitlab: Add issue template
It is common to see many issues without a proper bug description which
requires unnecessary cycles of questions and answers. This commit tries
to take advantage of a Gitlab feature named `issue template` by adding a
template for issue creation. This feature will be visible in a field
named "template" available in the process of devising a new ticket.
Unfortunately looks like Gitlab does not support setting a default issue
template which means that we still rely on users selecting this template
during the issue creation.

Signed-off-by: Rodrigo Siqueira <Rodrigo.Siqueira@amd.com>